Docker/geojson-app/requirements.txt
boto3==1.35.18
geojson==3.1.0
psycopg2-binary==2.9.9